+++ website/docs/components/select/examples/options.tsx
@@ -26,11 +26,18 @@ const Demo = () => (
           I'm disabled option-checkbox
         </Select.Option>
         <Select.Option value={3}>
-          <Select.Option.Checkbox indeterminate />
-          I'm indeterminate option-checkbox
+          <Select.Option.Content>
+            <Select.Option.Checkbox indeterminate />
+            I'm indeterminate option-checkbox
+          </Select.Option.Content>
+          <Select.Option.Hint>I'm hint</Select.Option.Hint>
         </Select.Option>
-        <Select.OptionTitle>I'm title</Select.OptionTitle>
-        <Select.OptionHint>I'm hint</Select.OptionHint>
+
+        <Select.Group title={"I'm title"} subTitle={"I'm hint"}>
+          <Select.Option value={4}>Option 1</Select.Option>
+          <Select.Option value={5}>Option 2</Select.Option>
+          <Select.Option value={6}>Option 3</Select.Option>
+        </Select.Group>
       </Select.Menu>
     </Select>
   </Flex>