+++ website/docs/components/select/examples/options_filtering.tsx
@@ -26,7 +26,7 @@ const Demo = () => {
         Options filtering
       </Text>
       <Select placeholder='Select value'>
-        <Select.Trigger id='options-filtering-select' mr='auto' mt={2} aria-controls={undefined} />
+        <Select.Trigger id='options-filtering-select' mr='auto' mt={2} />
         <Select.Popper>
           {({ highlightedIndex }) => (
             <>